+++ ImmigrantLife/Assets/_scriptables/_dialogues/oi yeah.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5d43fd573f6f2e049a9bab545a4a6743, type: 3}
+  m_Name: oi yeah
+  m_EditorClassIdentifier: 
+  dialogueList:
+  - Speaker: {fileID: 11400000, guid: fbe522b74a378744ab8290dbe123a485, type: 2}
+    Sentence: oh yeah
