+++ ImmigrantLife/Assets/_scriptables/_dialogues/oi yeah.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b5f34c8b5d084242904178fd7dcd1b3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
