+++ ImmigrantLife/Assets/_scriptables/_events.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 31ec0e57f3243684ebf90a85d6ca0a36
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
