+++ ImmigrantLife/Assets/_scriptables/_events/Eveent1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 637cc281266e37c4fbb1558b78879fd0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
