--- ImmigrantLife/Assets/_scripts/CharEnum.cs
@@ -1,21 +1,0 @@
-using UnityEngine;
-
-public class CharEnum : MonoBehaviour
-{
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
-public enum characters
-{
-    char1,
-    char2,
-}
