--- ImmigrantLife/Assets/_scripts/CharEnum.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c9ad78b911f802c49aa5c5391ce3c669
