--- ImmigrantLife/Assets/_scripts/CharacterScriptable.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-[CreateAssetMenu(fileName = "CharacterScriptable", menuName = "Scriptable Objects/CharacterScriptable")]
-public class CharacterScriptable : ScriptableObject
-{
-//char info
-    public string name;
-}
