--- ImmigrantLife/Assets/_scripts/CharacterScriptable.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ee2e26d45d2483c4b988962016bbd2bd
