--- ImmigrantLife/Assets/_scripts/DialogueManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1a03e3508d631f945b0abe6b07c3c323
