--- ImmigrantLife/Assets/_scripts/DialogueReader.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 118bce778040b2a49ba12fdc630e5581
