+++ ImmigrantLife/Assets/_scripts/_Managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 632592ab03b9327438e97752beb4958b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
