+++ ImmigrantLife/Assets/_scripts/_Managers/DialogueManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1a03e3508d631f945b0abe6b07c3c323
