+++ ImmigrantLife/Assets/_scripts/_Managers/DialogueReader.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 118bce778040b2a49ba12fdc630e5581
