+++ ImmigrantLife/Assets/_scripts/_Managers/EventManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5dbc9aa9c4f05c4419a2e16de57ac619
