+++ ImmigrantLife/Assets/_scripts/_scriptables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 65fcda56939a2db429ee34bf355a834b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
