+++ ImmigrantLife/Assets/_scripts/_scriptables/CharacterScriptable.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "CharacterScriptable", menuName = "Scriptable Objects/CharacterScriptable")]
+public class CharacterScriptable : ScriptableObject
+{
+//char info
+    public string name;
+}
