+++ ImmigrantLife/Assets/_scripts/_scriptables/CharacterScriptable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ee2e26d45d2483c4b988962016bbd2bd
