+++ ImmigrantLife/Assets/_scripts/_scriptables/EventDialogue.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dcc5bf02536c1d44b9b72b4e2d1e7356
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
