+++ ImmigrantLife/Assets/_scripts/_scriptables/EventDialogue.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2afce3a6399db004bb87347e745a839e
