+++ ImmigrantLife/Assets/_scripts/_scriptables/scriptableDialogue.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5d43fd573f6f2e049a9bab545a4a6743
