--- ImmigrantLife/Assets/_scripts/scriptableDialogue.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5d43fd573f6f2e049a9bab545a4a6743
