--- ImmigrantLife/Logs/shadercompiler-AssetImportWorker1-0.log
@@ -1,3 +1,0 @@
-Base path: 'C:/Program Files/Unity/Hub/Editor/6000.0.28f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/6000.0.28f1/Editor/Data/PlaybackEngines'
-Cmd: initializeCompiler
-
